feat(logging): Implement debug logging functionality
- Added a logging mechanism to output debug information to both console and a log file.
- Configured logging to capture unhandled exceptions and log the working directory.

refactor(track): Enhance UTrackPoint loading and error handling

- Introduced token trimming and sanitization for loading track points.
- Improved error handling during parsing of position and handle coordinates.

fix(ui): Prevent buffer data calls with empty vectors

- Updated buffer data calls in UPathRenderer to handle empty points and circles gracefully.

fix(viewport): Improve viewport resizing and rendering logic

- Added checks to prevent unnecessary framebuffer recreation.
- Enhanced viewport rendering with additional UI elements for grid and node legends.

fix(picker): Add error handling and lazy initialization for viewport picker

- Implemented error handling during shader creation and loading.
- Added lazy initialization for the viewport picker to avoid unnecessary resource allocation.

fix(uiutil): Ensure proper width handling in text input rendering

- Cast width to float when pushing item width in UI utility functions.

chore(todos): Update TODO list with current priorities and completed tasks

- Reorganized TODO items into categories and marked completed tasks.

+++ src/util/uiutil.cpp
@@ -15,7 +15,7 @@ int TextInputCallback(ImGuiInputTextCallbackData* data)
 
 bool UIUtil::RenderTextInput(std::string name, std::string* value, const int width)
 {
-	ImGui::PushItemWidth(width);
+	ImGui::PushItemWidth(static_cast<float>(width));
 
 	bool result = ImGui::InputText(name.c_str(), value->data(), value->size() + 1, ImGuiInputTextFlags_CallbackResize, TextInputCallback, value);
